+++ src/wamytmapp/tests/__init__.py
@@ -1,0 +1,5 @@
+from .addtimetangeform import *
+from .oudelegate import *
+from .selectlistitems import *
+from .timerangemanager import *
+from .views import *
